FastAPISocketIO/app/requirements.txt
fastapi
uvicorn
python-socketio[client]
python-socketio